Bug fixing #2
This commit is to address partially issue #2, now it can show how many messages when filter is selected and search query is empty. The format only shows the numbers but not the same with the mock up. The next patch will brings in the numbers in right format and clickable with the arrow button up and down.